OSP-5: Move repositories up one level
