+++ shyft/repository/raster_cell_file_repository.py
