+++ shyft/repository/smg_data.py
